Fixed datatype in one more position

+++ AnimeGAN/telegram/datasphere/request.py
@@ -17,7 +17,7 @@ def make_request(node_id: str, folder_id: str, model_id: str, model_input, datat
     input_shape = model_input.shape
     payload = httpclient.InferInput("input_variable_0", input_shape, datatype)
 
-    input_batch = np.array(model_input, dtype=np.float32)
+    input_batch = np.array(model_input)
     payload.set_data_from_numpy(input_batch, binary_data=False)
 
     inputs = [payload]
